MD2 Loading
Finished the MD2 Loader which was listed in the last commit. There is a
problem at the moment, however. and it appears to be with calling
functions to do with the Direct3D device from the ObjectModel class. My
C++ knowledge isn't good enough to fix it. Maybe it's using namespace?


Former-commit-id: 5fb0278c501120cb46631d22f3c8d4848c9caa32