Added 3D Model File Templates, other
Added some 3D Model File Template Headers for .MS3D and the custom .MSM,
also updated the .MD2 type one so we don't define types twice by
accident.

There was also an update to the .gitignore files to finish with the
previous update to VS2015.

A cleanup of junk files was also done.