Added Editor
You'll find it in /Tools/Editor/. It's very unfinished, to the point
it's unusable. It wouldn't be on here yet if it wasn't for the fact I've
got to reinstall Windows, and don't have anywhere safe to keep it while
I do so.

A few other small things inside the main engine have also been done, but
nothing of importance.