More Particles
Okay, the rendering and movement isn't done yet, but the actual particle
system and emitter is. I've also loosely added this to the Game Maker
project to test real-time speed of particle buffer operations.

There's still a lot I want to change and do, but it seems like good
progress so far.


Former-commit-id: dffaa986a22bc819b0cb1c7d0068c361d1457845