Work on Model System #1
Added MSM, a lightweight 3D geometry container to be used alongside MPM.
Added a struct for animation data.

Removed the Documentation folder from the repository, it housed outdated
formats.

Fixed a mistake with the Load() declaration inside CModelLoadMS3D.

Note:
Not much, but I'm busy with work at the moment.

Former-commit-id: 3b899388b99a732107c56afdc50050868d5d602f